+++ lumenairy/elements/__init__.py
@@ -88,6 +88,7 @@ from .lenses import (
     surface_sag_biconic,
     surface_sag_general,
 )
+from .pmm import pmm_efficiency_1d
 from .polarization import (
     JonesField,
     apply_half_wave_plate,
@@ -160,6 +161,7 @@ __all__ = [
     # RCWA / Fourier Modal Method (rigorous vector grating solver)
     'rcwa_efficiency_1d',
     'rcwa_efficiency_1d_jax',
+    'pmm_efficiency_1d',
     'rcwa_efficiency_2d',
     'rcwa_efficiency_2d_shapes',
     'rcwa_efficiency_vs_wavelength',